Added support for IPv6.
We have added an <IP> tag that is voluntarily lose and that should be
able to catch both IPv4 and IPv6 addresses (and maybe more).

The validity of the caught IP address is then checked in Python rather
than through a regex.

For now the validity check is done in `rig_actions/ipset.py`. That might
not be the best idea.

Also note that we now need two ipset, one to store IPv4 addresses (that
MUST be called *rig_blacklist4*) and another one to store IPv6
addresses (that MUST be called *rig_blacklist6*).